implemented multiselect for tags, removed form to own directroy
